pull out vertical bar chart component
