--- ui/app/components/vertical-bar-chart.js
@@ -1,16 +1,0 @@
-import Component from '@glimmer/component';
-
-/**
- * @module VerticalBarChart
- * VerticalBarChart components are used to...
- *
- * @example
- * ```js
- * <VerticalBarChart @requiredParam={requiredParam} @optionalParam={optionalParam} @param1={{param1}}/>
- * ```
- * @param {object} requiredParam - requiredParam is...
- * @param {string} [optionalParam] - optionalParam is...
- * @param {string} [param1=defaultValue] - param1 is...
- */
-
-export default class VerticalBarChart extends Component {}
