+++ ui/app/templates/components/clients/dashboard.hbs
@@ -160,16 +160,18 @@
         <Clients::TotalClientUsage
           @title="Vault usage"
           @description="This data can be used to understand how many total clients are using Vault each month for the time period selected above."
+          @chartLegend={{this.chartLegend}}
           @subTitle="Total monthly clients"
           @subText="Each unique client is counted once per month. This can help with capacity planning."
-          {{!-- yield bar chart component?!  --}}
-          @dataset={{this.verticalBarChartData}}
-          @chartLegend={{this.chartLegend}}
           @dataOne="Average total clients per month"
           @dataOneData="100"
           @dataTwo="Average new clients per month"
           @dataTwoData="4"
-        />
+        >
+          <Clients::VerticalBarChart
+            @dataset={{this.verticalBarChartData}}
+          />
+        </Clients::TotalClientUsage>
           {{!-- ARG TODO this is the search select from the old graph that can be reused --}}
           {{!-- <div class="column">
             <div class="card">
