+++ ui/app/templates/components/clients/total-client-usage.hbs
@@ -5,10 +5,9 @@
       <p class="chart-description">{{@description}}</p>
     {{/if}}
   </div>
-  <div class="chart-container-wide" {{on "mouseleave" (fn this.removeTooltip)}}>
-    <svg class="chart has-grid"
-      {{did-insert this.registerListener @dataset}}
-    ></svg>
+ 
+  <div class="chart-container-wide">
+    {{yield}}
   </div>
 
   <div class="chart-subTitle">
@@ -35,20 +34,4 @@
     <span class="dark-dot"></span><span class="legend-label">{{capitalize @chartLegend.1.label}}</span>
   </div>
 
-  {{!-- TOOLTIP --}}
-  
-  {{#if this.tooltipTarget}}  
-  {{!-- Required to set tag name = div https://github.com/yapplabs/ember-modal-dialog/issues/290 --}}
-  {{!-- Component must be in curly bracket notation --}}
-    {{#modal-dialog
-      tagName='div'
-      tetherTarget=this.tooltipTarget
-      targetAttachment='bottom middle'
-      attachment='bottom middle'
-      offset='40px 0'
-    }}
-      <p class="chart-tooltip">{{this.hoveredLabel}}</p>
-      <div class="chart-tooltip-arrow" />
-    {{/modal-dialog}}
-  {{/if}}
 </div>
