--- ui/app/templates/components/vertical-bar-chart.hbs
@@ -1,1 +1,0 @@
-{{yield}}
